Assert equal numbers of Start/End internal control flow labels
If the number of Start internal control flow labels does not equal the
number of End internal control flow labels, the result might be subtle
intermittent problems related to register allocation.

This change adds an assertion that, at the completion of the Register
Allocation Phase, the register allocator must not indicate that it's
still nested within a region of internal control flow nor popped out of
too many regions of internal control flow.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>